fix balance-settings-item relocate
