+++ src/widgets/balance-settings/index.tsx
@@ -2,8 +2,8 @@ import classnames from 'classnames';
 import styles from './styles.module.css';
 import { SettingsIcon } from 'shared/ui/icons/settings-icon';
 import { ChangeEvent } from 'react';
-import BalanceSettingsItem from './balance-settings-item';
 import { Button } from 'shared/ui/button';
+import BalanceSettingsItem from './components/balance-settings-item';
 
 interface BalanceSettingsProps {
   extClassName?: string;
